# packages/contracts/openapi/arthalens-api.yaml
openapi: "3.1.0"
info:
  title: ArthaLens API
  version: "1.0.0"
  description: |
    ArthaLens REST API — India macroeconomic intelligence platform.

    **Data policy**: Official statistics are reproduced with full provenance metadata.
    Derived analytics, ML forecasts, and AI-generated interpretations are clearly
    distinguished from official government estimates and are never presented as such.

    All monetary values in INR Crore unless stated otherwise.
    All timestamps are UTC ISO-8601.
    All periods follow the format defined in the period_type/period_label schema.
  contact:
    name: ArthaLens
  license:
    name: See LICENSE
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /gdp/revisions:
    get:
      tags: [gdp, revisions]
      operationId: getGdpRevisions
      summary: GDP revision history
      parameters:
        - $ref: "#/components/parameters/BaseYearParam"
        - $ref: "#/components/parameters/PeriodParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: Revision history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionsResponse"
...
  /ratings:
    get:
      tags: [ratings]
      operationId: getRatings
      summary: Sovereign credit ratings for India
      parameters:
        - name: agency
          in: query
          schema:
            type: string
            enum: [moodys, sp, fitch, all]
            default: all
      responses:
        "200":
          description: Ratings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingsResponse"
...
components:
  parameters:
    BaseYearParam:
      name: base_year
      in: query
      schema:
        type: string
        enum: ["2011-12", "2022-23"]
        default: "2022-23"
      description: GDP base year series to use
...
    PeriodParam:
      name: period
      in: query
      schema:
        type: string
        example: "2023-24"
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    Provenance:
      type: object
      description: Source and methodology provenance attached to every observation
      properties:
        sourceId:
          type: string
          format: uuid
        sourceName:
          type: string
        sourceUrl:
          type: string
          format: uri
        publicationDate:
          type: string
          format: date
        retrievalTimestamp:
          type: string
          format: date-time
        baseYear:
          type: string
          example: "2022-23"
        methodologyVersion:
          type: string
        observationStatus:
          type: string
          enum: [official, provisional, revised, estimated, derived, forecast]
...
    RevisionsResponse:
      type: object
      properties:
        period:
          type: string
        baseYear:
          type: string
        revisions:
          type: array
          items:
            type: object
            properties:
              revisionDate:
                type: string
                format: date
              fromValue:
                type: number
              toValue:
                type: number
              fromStatus:
                type: string
              toStatus:
                type: string
              releaseLabel:
                type: string
              provenance:
                $ref: "#/components/schemas/Provenance"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    RatingsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              agency:
                type: string
              rating:
                type: string
              outlook:
                type: string
                enum: [Positive, Stable, Negative, Watch]
              ratingDate:
                type: string
                format: date
              previousRating:
                type: string
                nullable: true
              actionType:
                type: string
                enum: [Upgrade, Downgrade, Affirm, Outlook_Change]
              provenance:
                $ref: "#/components/schemas/Provenance"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer